party: use separate lockfile.
Make Party use a separate lock file in a protected
directory that normal users can't get access to.  This
prevents the sort of abuse where a user might open the
channel log and then take a lock on it, thus preventing
users from appending to the party log.

Now, they may still do so but it's irrelevant, since
the channel log itself is no longe rpart of the channel
locking protocol on writes.

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>