web: replace old php pages with xhtml5
This was an automated conversion.  The version of PHP
used by M-Net was too old to be understood by the current
interpreters, so I ran it on the old M-Net to generate
HTML, which I then copied here.  That resulted in a number
of crufty HTML files, so I then ran those through tidy5
and xmlformat, to get well-formed XML.

Next, I updated internal links so that they refer to the
HTML pages, not PHP, via sed:

    sed -i '' 's/href="\([-a-z0-9_]*\)\.php/href="\1.html/' *.html

Finally, I took the CSS that had been at the top of each
file and moved it into a separate file under the `style/`
subdirectory.

What's here is imperfect, but a reasonable starting point
for cleaning up M-Net's web site.

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>